workflow: add action workflow to mirror OpenHarmony from gitee
This workflow sets up a mirror job with the hub-mirror-action.
It loops over all ~670  repositories in the OpenHarmony gitee
organization and mirrors them one-way to this account.
The job is triggered once every 24h via a schedule, as well as
by a push to the main branch of the repo.

Due to size and time requirements this job needs a self-hosted
runner. The job curenrtly has a timeout value of 23h. Speed of
mirroring does depend on network connection and transit to China.
In our test setup the job finishes after ~19h.

Out of the ~670 repos, 14 are difficult to mirror at the moment (repo
size, lfs usage with files exceeding GitHub policy, etc). For now we
exclude these. Work is in progress to reduce this list.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>