main.yml: spread out sync into 8 parallel jobs
This helps to speed up the whole sync process and stays within the 6h
timeout for runners. An initial job is used to gather the current list
of repos. To only retrieve the repo list instead of doing a full sync
we are using our own modified fork of the hub-mirror-action for now.

A small python script is used to handle splitting up the repo list in
chunks and also doing some processing to remove known-problematic repos.
They will be back once we have the issues around them debugged and
fixed.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>